add auth

+++ dashboard/app/db/categories/index.ts
@@ -1,25 +1,33 @@
 "use server";
 
+import { getSession } from "@/lib/auth";
 import uri from "@/lib/uri";
-import { revalidateTag, unstable_cache } from "next/cache";
+import { revalidateTag, unstable_cache, unstable_noStore } from "next/cache";
 
-export const getCategories = unstable_cache(
-  async () => {
-    try {
-      const res = await fetch(`${uri}/categories`);
-      if (!res.ok) {
-        return [];
-      }
-      const data: { data: Category[] } = await res.json();
-      return data.data;
-    } catch (error) {
-      console.log(error);
+export const getCategories = async () => {
+  unstable_noStore();
+
+  try {
+    const authUser = await getSession();
+    if (!authUser) {
       return [];
     }
-  },
-  ["categories"],
-  { tags: ["categories"] }
-);
+    const res = await fetch(`${uri}/categories`, {
+      headers: {
+        "Content-Type": "application/json",
+        Authorization: `Bearer ${authUser.token}`,
+      },
+    });
+    if (!res.ok) {
+      return [];
+    }
+    const data: { data: Category[] } = await res.json();
+    return data.data;
+  } catch (error) {
+    console.log(error);
+    return [];
+  }
+};
 
 export const createCategory = async ({
   categroy,
@@ -27,11 +35,16 @@ export const createCategory = async ({
   categroy: Omit<Category, "id" | "brandIds" | "productIDs">;
 }): Promise<{ message: string }> => {
   try {
+    const authUser = await getSession();
+    if (!authUser) {
+      return { message: "يجب تسجيل الدخول للتمكن من اكمال العملية" };
+    }
     const res = await fetch(`${uri}/categories`, {
       method: "POST",
       body: JSON.stringify(categroy),
       headers: {
-        "Content-Type": "application/json", // Add the Content-Type header
+        "Content-Type": "application/json",
+        Authorization: `Bearer ${authUser.token}`,
       },
     });
 
@@ -54,11 +67,16 @@ export const updateCategory = async ({
   categroy: Omit<Category, "brandIds" | "productIDs">;
 }): Promise<{ message: string }> => {
   try {
+    const authUser = await getSession();
+    if (!authUser) {
+      return { message: "يجب تسجيل الدخول للتمكن من اكمال العملية" };
+    }
     const res = await fetch(`${uri}/categories/${categroy.id}`, {
       method: "PUT",
       body: JSON.stringify(categroy),
       headers: {
-        "Content-Type": "application/json", // Add the Content-Type header
+        "Content-Type": "application/json",
+        Authorization: `Bearer ${authUser.token}`,
       },
     });
 
@@ -81,11 +99,16 @@ export const deleteCategory = async ({
   id: string;
 }): Promise<{ message: string }> => {
   try {
+    const authUser = await getSession();
+    if (!authUser) {
+      return { message: "يجب تسجيل الدخول للتمكن من اكمال العملية" };
+    }
     const res = await fetch(`${uri}/categories/${id}`, {
       method: "DELETE",
       // body: JSON.stringify(categroy),
       headers: {
-        "Content-Type": "application/json", // Add the Content-Type header
+        "Content-Type": "application/json",
+        Authorization: `Bearer ${authUser.token}`,
       },
     });
 
@@ -108,11 +131,16 @@ export const updateBrand = async ({
   brand: Omit<Brand, "productIDs">;
 }): Promise<{ message: string }> => {
   try {
+    const authUser = await getSession();
+    if (!authUser) {
+      return { message: "يجب تسجيل الدخول للتمكن من اكمال العملية" };
+    }
     const res = await fetch(`${uri}/brands/${brand.id}`, {
       method: "PUT",
       body: JSON.stringify(brand),
       headers: {
-        "Content-Type": "application/json", // Add the Content-Type header
+        "Content-Type": "application/json",
+        Authorization: `Bearer ${authUser.token}`,
       },
     });
 
@@ -135,10 +163,15 @@ export const deleteBrand = async ({
   id: string;
 }): Promise<{ message: string }> => {
   try {
+    const authUser = await getSession();
+    if (!authUser) {
+      return { message: "يجب تسجيل الدخول للتمكن من اكمال العملية" };
+    }
     const res = await fetch(`${uri}/brands/${id}`, {
       method: "Delete",
       headers: {
-        "Content-Type": "application/json", // Add the Content-Type header
+        "Content-Type": "application/json",
+        Authorization: `Bearer ${authUser.token}`,
       },
     });
 
@@ -161,11 +194,16 @@ export const createBrand = async ({
   brand: Omit<Brand, "id" | "productIDs">;
 }): Promise<{ message: string }> => {
   try {
+    const authUser = await getSession();
+    if (!authUser) {
+      return { message: "يجب تسجيل الدخول للتمكن من اكمال العملية" };
+    }
     const res = await fetch(`${uri}/brands`, {
       method: "POST",
       body: JSON.stringify(brand),
       headers: {
-        "Content-Type": "application/json", // Add the Content-Type header
+        "Content-Type": "application/json",
+        Authorization: `Bearer ${authUser.token}`,
       },
     });
 
@@ -183,20 +221,26 @@ export const createBrand = async ({
   }
 };
 
-export const getBrands = unstable_cache(
-  async () => {
-    try {
-      const res = await fetch(`${uri}/brands`);
-      if (!res.ok) {
-        return [];
-      }
-      const data: { data: Brand[] } = await res.json();
-      return data.data;
-    } catch (error) {
-      console.log(error);
+export const getBrands = async () => {
+  unstable_noStore();
+  try {
+    const authUser = await getSession();
+    if (!authUser) {
       return [];
     }
-  },
-  ["categories"],
-  { tags: ["categories"] }
-);
+    const res = await fetch(`${uri}/brands`, {
+      headers: {
+        "Content-Type": "application/json",
+        Authorization: `Bearer ${authUser.token}`,
+      },
+    });
+    if (!res.ok) {
+      return [];
+    }
+    const data: { data: Brand[] } = await res.json();
+    return data.data;
+  } catch (error) {
+    console.log(error);
+    return [];
+  }
+};
